removing IPublisherRepository
