+++ BookStore.Domain/BookstoreApp.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BookStore.Domain
+{
+    public class BookstoreApp
+    {
+
+    }
+}
