+++ BookStore.Domain/Models/IBookStoreDb.cs
@@ -1,9 +1,13 @@
-﻿namespace BookStore.Domain.Models
+﻿using System.Collections.Generic;
+
+namespace BookStore.Domain.Models
 {
     public interface IBookStoreDb
     {
         IRepository<Book> Books { get; }
 
-        IPublisherRepository Publishers { get; }
+        IRepository<Publisher> Publishers { get; }
+
+        IEnumerable<Book> GetPublisherBooks(Publisher publisher);
     }
 }
