--- BookStore.Domain/Models/IPublisherRepository.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace BookStore.Domain.Models
-{
-    public interface IPublisherRepository : IRepository<Publisher>
-    {
-        IEnumerable<Book> GetPublisherBooks(Publisher publisher);
-    }
-}
