+++ BookStore/Repositories/PublisherRepository.cs
@@ -6,7 +6,7 @@ using System.Linq;
 
 namespace BookStore.Repositories
 {
-    public class PublisherRepository : RepositoryBase<Publisher>, IPublisherRepository
+    public class PublisherRepository : RepositoryBase<Publisher>
     {
         public PublisherRepository(DbConnection connection, IBookStoreDb parent) 
             : base(connection, parent)
@@ -34,18 +34,6 @@ namespace BookStore.Repositories
             {
                 return false;
             }
-        }
-
-        public IEnumerable<Book> GetPublisherBooks(Publisher publisher)
-        {
-            IEnumerable<Book> books = Enumerable.Empty<Book>();
-
-            if (publisher != null)
-            {
-                books = Parent.Books.All().Where(book => book.Publisher.Id == publisher.Id);
-            }
-
-            return books;
         }
     }
 }
